Create mangrove-agent
A mangrove agent is defined in space by the x and y position of the center of its trunk. It is characterized by its species, which gives the species-specific parameters such as growth rate, allometry, wood density, and responses to environmental conditions. The state of a mangrove agent is defined by its DBH (Diameter at breast height), height, crown radius, above-ground biomass, and maturity state. A mangrove agent also has a memory of its growth for the last five years, which affects its mortality.